requirements.txt
discord.py
python-dotenv
yt-dlp
PyNaCl
aiohttp
fastapi
uvicorn
spotipy
